Seems as we do not have sboot source, here are some hacks to build with the bundled binaries
kconfig has been provided as a binary (32-bit Linux) only. So
building on OS X is not presently possible. KBuild is fairly
standard software though, so we can probably replace this version
later.

Also some scripts make mention to a crc binary that isn't included.
This actually doesn't matter because CRC checking has been disabled
in the KkApp and as such binaries *shouldn't* include a CRC.